+++ core/templates/services/request-interceptor.service.spec.ts
@@ -21,49 +21,11 @@ import {
   HttpTestingController
 } from '@angular/common/http/testing';
 // eslint-disable-next-line oppia/disallow-httpclient
-import {HTTP_INTERCEPTORS, HttpClient, HttpHandler, HttpRequest, HttpParams, HttpBackend} from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClient, HttpHandler, HttpRequest, HttpParams } from '@angular/common/http';
 
-import { RequestInterceptor } from 'services/request-interceptor.service';
+import { MockCsrfTokenService, RequestInterceptor } from 'services/request-interceptor.service';
 import { CsrfTokenService } from './csrf-token.service';
-import {Injectable} from "@angular/core";
 
-
-@Injectable({
-  providedIn: 'root'
-})
-export class MockCsrfTokenService {
-  // 'tokenPromise' will be null when token is not initialized.
-  tokenPromise: PromiseLike<string> | null = null;
-  http: HttpClient;
-
-  constructor(httpBackend: HttpBackend) {
-    this.http = new HttpClient(httpBackend);
-  }
-
-  initializeToken(): void {
-    if (this.tokenPromise !== null) {
-      throw new Error('Token request has already been made');
-    }
-    this.tokenPromise = this.http.get(
-      '/csrfhandler', { responseType: 'text' }
-    ).toPromise().then((responseText: string) => {
-      // Remove the protective XSSI (cross-site scripting inclusion) prefix.
-      return JSON.parse(responseText.substring(5)).token;
-    }, (err) => {
-      console.error(
-        'The following error is thrown while trying to get CSRF token.');
-      console.error(err);
-      throw err;
-    });
-  }
-
-  getTokenAsync(): PromiseLike<string> {
-    if (this.tokenPromise === null) {
-      throw new Error('Token needs to be initialized');
-    }
-    return this.tokenPromise;
-  }
-}
 
 describe('Request Interceptor Service', () => {
   let cts: CsrfTokenService;
@@ -92,18 +54,6 @@ describe('Request Interceptor Service', () => {
     // we need to mock the `getTokenAsync` function for testing purposes.
     // @ts-expect-error
     spyOn(cts, 'getTokenAsync').and.returnValue(['sample-csrf-token']);
-    // This throws "Argument of type '(options: any)' is not
-    // assignable to parameter of type 'jqXHR<any>'.". We need to suppress
-    // this error because we need to mock $.ajax to this function for
-    // testing purposes.
-    // @ts-expect-error
-    spyOn($, 'ajax').and.callFake((options: Promise) => {
-      let d = $.Deferred();
-      d.resolve(
-        options.dataFilter('12345{"token": "sample-csrf-token"}')
-      );
-      return d.promise();
-    });
   });
 
   afterEach(() => {
@@ -115,9 +65,12 @@ describe('Request Interceptor Service', () => {
       async(response) => expect(response).toBeTruthy());
 
     let req = httpTestingController.expectOne('/api');
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
     expect(req.request.method).toEqual('POST');
     expect(req.request.body.constructor.name).toEqual('FormData');
+    expect(reqCSFR.request.method).toEqual('GET');
     req.flush({data: 'test'});
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
   });
 
   it('should modify http requests body when they are intercepted', () => {
@@ -125,8 +78,11 @@ describe('Request Interceptor Service', () => {
       async(response) => expect(response).toBeTruthy());
 
     let req = httpTestingController.expectOne('/api');
+    let reqCSRF = httpTestingController.expectOne('/csrfhandler');
     expect(req.request.method).toEqual('PATCH');
+    expect(reqCSRF.request.method).toEqual('GET');
     req.flush({data: 'test'});
+    reqCSRF.flush('12345{"token": "sample-csrf-token"}');
 
     expect(req.request.body).toEqual({
       csrf_token: 'sample-csrf-token',
@@ -143,10 +99,13 @@ describe('Request Interceptor Service', () => {
 
     let req = httpTestingController.expectOne('/api');
     let req2 = httpTestingController.expectOne('/api2');
+    let reqCSRF = httpTestingController.expectOne('/csrfhandler');
     expect(req.request.method).toEqual('GET');
     expect(req2.request.method).toEqual('PATCH');
+    expect(reqCSRF.request.method).toEqual('GET');
     req.flush({data: 'test'});
     req2.flush({data: 'test'});
+    reqCSRF.flush('12345{"token": "sample-csrf-token"}');
 
     expect(cts.getTokenAsync).toHaveBeenCalledTimes(1);
     expect(req.request.body).toBeNull();
@@ -174,10 +133,16 @@ describe('Request Interceptor Service', () => {
     mockCsrfTokenService.getTokenAsync().then(function(token) {
       expect(token).toEqual('sample-csrf-token');
     }).then(done, done.fail);
+
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
   });
 
   it('should error if initialize is called more than once', () => {
     mockCsrfTokenService.initializeToken();
+
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
 
     expect(() => mockCsrfTokenService.initializeToken())
       .toThrowError('Token request has already been made');
@@ -221,6 +186,9 @@ describe('Request Interceptor Service', () => {
           {params: new HttpParams({fromString: 'key=null'})}),
           {} as HttpHandler);
     }).toThrowError('Cannot supply params with value "None" or "null".');
+
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
   });
 
   it('should throw error if param with None value is supplied', () => {
@@ -232,6 +200,9 @@ describe('Request Interceptor Service', () => {
           {params: new HttpParams({fromString: 'key=None'})}),
           {} as HttpHandler);
     }).toThrowError('Cannot supply params with value "None" or "null".');
+
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
   });
 
   it('should throw error if null param in DELETE request', () => {
@@ -243,6 +214,9 @@ describe('Request Interceptor Service', () => {
           {params: new HttpParams({fromString: 'key=null'})}),
           {} as HttpHandler);
     }).toThrowError('Cannot supply params with value "None" or "null".');
+
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
   });
 
   it('should not throw error if null param in POST request', () => {
@@ -254,6 +228,8 @@ describe('Request Interceptor Service', () => {
 
     const req = httpTestingController.expectOne('/api?key=null');
     req.flush({data: 'test'});
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
   });
 
   it('should not throw error if null param in PUT request', () => {
@@ -265,6 +241,8 @@ describe('Request Interceptor Service', () => {
 
     const req = httpTestingController.expectOne('/api?key=null');
     req.flush({data: 'test'});
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
   });
 
   it('should not throw error if null param in PATCH request', () => {
@@ -276,5 +254,7 @@ describe('Request Interceptor Service', () => {
 
     const req = httpTestingController.expectOne('/api?key=null');
     req.flush({data: 'test'});
+    let reqCSFR = httpTestingController.expectOne('/csrfhandler');
+    reqCSFR.flush('12345{"token": "sample-csrf-token"}');
   });
 });
